Add basic rover instruction following
There is currently nothing to prevent the rover leaving the plateau.

There is currently nothing to take the input as a string, or to produce
the output as a string.

I have not directly tested some of the details (eg that rotating left
from North points you West) because these details are all covered by the
tests from the problem description.

I don't believe there is value in testing the details directly, as they
would mostly be repeating the implementation (eg the implementation says
that moving North increases your y co-ordinate and the test assert that
moving North increases your y co-ordinate).

In a more complicated problem, there might be value in property testing
the details (eg rotating left 4 times returns you to your original
heading). I think that would be overkill for this problem.

The tests are the tests from the problem description.