Produce interactive program
It works, but I'm not yet happy with it.

There is nothing to stop the robot leaving the plateau. Becuase I don't
yet use the plateau, clippy comments about the variable being unused.

It's a little bit messy. Some types know how to parse themselves, and
others don't. It could do with being made more consistent.

It feels longer than it ought to be. Possibly this is because parsing
code in Rust is fairly verbose, especially when different error states
are handled.