Crash rovers outside of the plateau
I'm currently doing the check within the main file.

This doesn't feel great.

I don't think the plateau should belong to the rover, or the rover to
the plateau.

I think I am missing a World or Simulation struct that would know about
the rover and the plateau and confirm the rover isn't out of bounds. I
may explore introducing this, but I suspect the extra complexity isn't
worth the purity of the solution.